Separating download from retrain
