+++ cnn/download_model.py
@@ -1,0 +1,70 @@
+import os.path
+import sys
+import tarfile
+import argparse
+
+from utils import create_model_info
+
+def maybe_download_and_extract(data_url, model_dir):
+  """Download and extract model tar file.
+
+  If the pretrained model we're using doesn't already exist, this function
+  downloads it from the TensorFlow.org website and unpacks it into a directory.
+
+  Args:
+    data_url: Web location of the tar file containing the pretrained model.
+    model_dir: Directory where to check if model exists and where to store
+        it if it does not exist.
+  """
+  dest_directory = model_dir
+  if not os.path.exists(dest_directory):
+    os.makedirs(dest_directory)
+  filename = data_url.split('/')[-1]
+  filepath = os.path.join(dest_directory, filename)
+  if not os.path.exists(filepath):
+
+    def _progress(count, block_size, total_size):
+      sys.stdout.write('\r>> Downloading %s %.1f%%' %
+                       (filename,
+                        float(count * block_size) / float(total_size) * 100.0))
+      sys.stdout.flush()
+
+    filepath, _ = urllib.request.urlretrieve(data_url, filepath, _progress)
+    print()
+    statinfo = os.stat(filepath)
+    tf.logging.info('Successfully downloaded', filename, statinfo.st_size,
+                    'bytes.')
+    print('Extracting file from ', filepath)
+    tarfile.open(filepath, 'r:gz').extractall(dest_directory)
+  else:
+    print('Not extracting or downloading files, model already present in disk')
+
+if __name__ == '__main__':
+  parser = argparse.ArgumentParser()
+  parser.add_argument(
+      '--model_dir',
+      type=str,
+      default='/tmp/imagenet',
+      help="""\
+      Path to classify_image_graph_def.pb,
+      imagenet_synset_to_human_label_map.txt, and
+      imagenet_2012_challenge_label_map_proto.pbtxt.\
+      """
+  )
+  parser.add_argument(
+      '--architecture',
+      type=str,
+      default='inception_v3',
+      help="""\
+      Which model architecture to use. 'inception_v3' is the most accurate, but
+      also the slowest. For faster or smaller models, chose a MobileNet with the
+      form 'mobilenet_<parameter size>_<input_size>[_quantized]'. For example,
+      'mobilenet_1.0_224' will pick a model that is 17 MB in size and takes 224
+      pixel input images, while 'mobilenet_0.25_128_quantized' will choose a much
+      less accurate, but smaller and faster network that's 920 KB on disk and
+      takes 128x128 images. See https://research.googleblog.com/2017/06/mobilenets-open-source-models-for.html
+      for more information on Mobilenet.\
+      """)
+  FLAGS, unparsed = parser.parse_known_args()
+  info = create_model_info(FLAGS.architecture)
+  maybe_download_and_extract(info["data_url"], FLAGS.model_dir)
